+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 - Feature: Add `avatarBadge` props on `Comment` and `CommentForm`.
+- Fix: Fix `width` error styles `TextInputWithUnit`.
 
 ## [2.13.1] - 2019-07-08